+++ src/main/java/com/bancomachado/bancomachado/controller/request/SavingsAccountRequest.java
@@ -1,4 +1,4 @@
 package com.bancomachado.bancomachado.controller.request;
 
-public record SavingsAccountRequest(int agency, int account, Long balance) {
+public record SavingsAccountRequest(int agency, int account, Double balance) {
 }
